Currently, 8x8 block size results have been produce.
The kernels need double checked to ensure that the times are aligning appropriately.
There is an issue when writing back to the shared array at times. That is the two
shapes are not the same. This may be due to the determination of the number of swept steps.

Start with the swept steps determination and work past the kernels to ensure that times
are lining up appropriately.
Also, check the octahedron step to ensure that It is not over calculating.